Introduce strict mode
This introduces a new parameter "strict" in order to specify whether or
not the permissions sync should continue in case errors are encountered.
When strict mode is enabled (strict=true), then all exceptions will be
thrown immediately. In situations where wildcard permissions are set, it
might be undesirable to let the permissions sync fail entirely, so
setting strict=false will skip any exceptions and continue the sync.